Create First_attempt
This is a first attempt. Check the code, see what you understand or not and you can ask me stuff if ever you don't understand or chatgpt (Its all my original code but I mean for sure it understands). I think we need to implement some kind of queue so that it loops back to unfinished processes. To make sure the process stops after the quantum time we have a condition.wait() that puts the thread to sleep. its commented out right now cause there's no .notify() which would wake it up so when we want to. I went backwards in the programming (meaning I started with process run, then process select, then user select) and I tested out the process select with actual values and it works. same for reading the file it shoots out the correct values. I think it's a decent start, obviously still a lot of stuff to do but I think we should built from this. Let me know what you think.